new file to handle initialization of networks
